refactor: 패키지 정리 및 공백 추가
- 패미지 정리. 테이블 나타내는 클래스 (Post, User)  db/entity로 이동
- 컨벤션 통일 위해서 클래스 첫줄, 마지막줄 공백 및 필드 사이 공백 추가
